+++ lib/rubocop/cop/style/copyright.rb
@@ -16,6 +16,9 @@ module RuboCop
       # an offense is reported.
       #
       class Copyright < Cop
+        AUTOCORRECT_EMPTY_WARNING = 'An AutocorrectNotice must be defined in' \
+                                    'your RuboCop config'.freeze
+
         def message
           "Include a copyright notice matching /#{notice}/" \
           'before any code.'
@@ -68,11 +71,12 @@ module RuboCop
         end
 
         def autocorrect(token)
-          raise Warning, 'An AutocorrectNotice must be defined in ' \
-            'your RuboCop config' if autocorrect_notice.empty?
+          raise Warning, AUTOCORRECT_EMPTY_WARNING if autocorrect_notice.empty?
           regex = Regexp.new(notice)
-          raise Warning, "AutocorrectNotice '#{autocorrect_notice}' must " \
-            "match Notice /#{notice}/" unless autocorrect_notice =~ regex
+          unless autocorrect_notice =~ regex
+            raise Warning, "AutocorrectNotice '#{autocorrect_notice}' must " \
+                           "match Notice /#{notice}/"
+          end
 
           lambda do |corrector|
             range = token.nil? ? range_between(0, 0) : token.pos